Not modifying config fixture since other tests might use it
